refactor(web): migrate legacy memory pages to app directory (#1)
Migrated `AgentMemoryGraphPage` and `MemoryDetailPage` from the legacy `pages/`
directory to the new `app/` directory following the `page.tsx` convention.
Also removed the redundant `ChannelsPage.tsx` from `pages/`.

- Renamed `web/src/pages/AgentMemoryGraphPage.tsx` -> `web/src/app/agents/_id/memory-graph/page.tsx`
- Renamed `web/src/pages/MemoryDetailPage.tsx` -> `web/src/app/memory/_id/page.tsx`
- Deleted `web/src/pages/ChannelsPage.tsx` (already present in `app/channels/page.tsx`)
- Updated routes in `web/src/main.tsx`
