+++ web/src/main.tsx
@@ -24,8 +24,8 @@ import SchedulesPage from '@/app/schedules/page';
 import SettingsPage from '@/app/settings/page';
 import ToolsPage from '@/app/tools/page';
 import MemoryExplorerPage from '@/app/memory/page';
-import MemoryDetailPage from '@/pages/MemoryDetailPage';
-import AgentMemoryGraphPage from '@/pages/AgentMemoryGraphPage';
+import MemoryDetailPage from '@/app/memory/_id/page';
+import AgentMemoryGraphPage from '@/app/agents/_id/memory-graph/page';
 import ChannelsPage from '@/app/channels/page';
 import TemplatesPage from '@/app/templates/page';
 import ProvidersPage from '@/app/providers/page';
